Blog/requirements.txt
asgiref==3.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-mysql==3.9.0
djangorestframework==3.12.2
mysqlclient==2.0.1
pika==1.1.0
pytz==2021.3
sqlparse==0.4.2
